Ligação da Agenda

+++ src/main/java/com/senac/beautyschedule/application/dto/ServicoRequest.java
@@ -1,9 +1,11 @@
 package com.senac.beautyschedule.application.dto;
 
+import java.math.BigDecimal;
+
 public record ServicoRequest (
         String descricao,
         double tempo,
-        double preco
+        BigDecimal preco
 ){
 
 }
